+++ baybe/recommenders/base.py
@@ -4,6 +4,7 @@ from typing import Protocol, runtime_checkable
 
 import cattrs
 import pandas as pd
+from cattrs import override
 
 from baybe.objectives.base import Objective
 from baybe.searchspace import SearchSpace
@@ -14,6 +15,10 @@ from baybe.serialization.core import get_base_structure_hook
 @runtime_checkable
 class RecommenderProtocol(Protocol):
     """Type protocol specifying the interface recommenders need to implement."""
+
+    # Use slots so that derived classes also remain slotted
+    # See also: https://www.attrs.org/en/stable/glossary.html#term-slotted-classes
+    __slots__ = ()
 
     def recommend(
         self,
@@ -47,15 +52,35 @@ class RecommenderProtocol(Protocol):
         ...
 
 
+# TODO: The workarounds below are currently required since the hooks created through
+#   `unstructure_base` and `get_base_structure_hook` do not reuse the hooks of the
+#   actual class, hence we cannot control things there. Fix is already planned and also
+#   needed for other reasons.
+
 # Register (un-)structure hooks
 converter.register_unstructure_hook(
     RecommenderProtocol,
     lambda x: unstructure_base(
         x,
         # TODO: Remove once deprecation got expired:
-        overrides=dict(acquisition_function_cls=cattrs.override(omit=True)),
+        overrides=dict(
+            acquisition_function_cls=cattrs.override(omit=True),
+            # Temporary workaround (see TODO note above)
+            _surrogate_model=override(rename="surrogate_model"),
+            _current_recommender=override(omit=False),
+            _used_recommender_ids=override(omit=False),
+        ),
     ),
 )
 converter.register_structure_hook(
-    RecommenderProtocol, get_base_structure_hook(RecommenderProtocol)
+    RecommenderProtocol,
+    get_base_structure_hook(
+        RecommenderProtocol,
+        # Temporary workaround (see TODO note above)
+        overrides=dict(
+            _surrogate_model=override(rename="surrogate_model"),
+            _current_recommender=override(omit=False),
+            _used_recommender_ids=override(omit=False),
+        ),
+    ),
 )